+++ jupyter/data/configs/electrical-grid-stability.json
@@ -1,0 +1,14 @@
+{
+    "job_id": "electrical-grid-stability",
+    "rows": 10000,
+    "cols": 13,  
+    "memory_limit": 50000000,
+    "threads": 3,
+    "results_field": "ml",
+    "analysis": {
+      "name": "regression",
+      "parameters": {
+        "dependent_variable": "stab"
+      }
+    }
+  }
