+++ jupyter/data/configs/ember_malware_bytes.json
@@ -1,0 +1,19 @@
+{
+    "job_id": "ember_malware_bytes",
+    "rows": 600000,
+    "cols": 258,
+    "memory_limit": 26843545600,
+    "threads": 10,
+    "results_field": "ml",
+    "categorical_fields": [
+        "label",
+        "sha256"
+    ],
+    "analysis": {
+        "name": "classification",
+        "parameters": {
+            "dependent_variable": "label",
+            "num_classes": 2
+        }
+    }
+}
