+++ jupyter/data/configs/regression_wind.json
@@ -1,0 +1,14 @@
+{
+    "job_id": "regression_wind",
+    "rows": 6574,
+    "cols": 15,  
+    "memory_limit": 50000000,
+    "threads": 3,
+    "results_field": "ml",
+    "analysis": {
+      "name": "regression",
+      "parameters": {
+        "dependent_variable": "MAL"
+      }
+    }
+  }
