+++ jupyter/data/configs/wine_quality_white.json
@@ -1,0 +1,14 @@
+{
+    "job_id": "wine_quality_white",
+    "rows": 4898,
+    "cols": 12,  
+    "memory_limit": 50000000,
+    "threads": 3,
+    "results_field": "ml",
+    "analysis": {
+      "name": "regression",
+      "parameters": {
+        "dependent_variable": "quality"
+      }
+    }
+  }
