--- jupyter/notebooks/configs/ccpp.json
@@ -1,14 +1,0 @@
-{
-    "job_id": "ccpp",
-    "rows": 9568,
-    "cols": 5,
-    "memory_limit": 50000000,
-    "threads": 3,
-    "results_field": "ml",
-    "analysis": {
-      "name": "regression",
-      "parameters": {
-        "dependent_variable": "PE"
-      }
-    }
-  }
