--- jupyter/notebooks/configs/regression_2175_kin8nm.json
@@ -1,14 +1,0 @@
-{
-    "job_id": "regression_2175_kin8nm",
-    "rows": 8192,
-    "cols": 9,   
-    "memory_limit": 50000000,
-    "threads": 3,
-    "results_field": "ml",
-    "analysis": {
-      "name": "regression",
-      "parameters": {
-        "dependent_variable": "y"
-      }
-    }
-  }
