--- jupyter/notebooks/configs/regression_house_16h.json
@@ -1,14 +1,0 @@
-{
-    "job_id": "regression_house_16h",
-    "rows": 22784,
-    "cols": 17,  
-    "memory_limit": 50000000,
-    "threads": 3,
-    "results_field": "ml",
-    "analysis": {
-      "name": "regression",
-      "parameters": {
-        "dependent_variable": "price"
-      }
-    }
-  }
