+++ jupyter/src/incremental_learning/config.py
@@ -1,0 +1,5 @@
+from pathlib2 import Path
+
+data_dir = Path('data')
+datasets_dir = data_dir / 'datasets'
+configs_dir = data_dir / 'configs'
