+++ jupyter/src/setup.py
@@ -1,0 +1,4 @@
+from setuptools import setup, find_packages
+
+setup(name="incremental_learning",
+      version="0.1", packages=find_packages())
